Squashed 'modules/pmi/' changes from 83433639f6..81076093b8
81076093b8 Improved test for disjoint selections.
ea98262769 fixed print function for python 3
266f6b7f6e creating a ThreeLetterCode to OneLetterCode class and testing it
a9ffa56913 check consistency of residue types between xl db and fasta
28dcc3ad4f aligning a short peptide sequence is not a good idea
c117aaf289 removing unused getRMSD class
d06c8bdace dict.iteritems() -> dict.items() for Python 3.6
2490719ff1 dict.keys() -> list(dict.keys()) for Python 3.6
f984d1a7b9 removing tabs
2416d85e81 fixed test
c2331fd690 added test data
c70c20046f added RMSD analysis
ba21ce0986 fixing COM calculation for RBs
ea64cd0e29 testing JaccardDistanceMatrix
1ee4b18f27 adding JaccardDistanceMatrix
4c2fc2e526 Fix handling of bool in Python 3.
b8c0facf6c Avoid syntax error in Python 3 with long literals.

git-subtree-dir: modules/pmi
git-subtree-split: 81076093b80a6a6d2ef0f0b5d9357428dee269b7

+++ modules/pmi/test/test_mmcif.py
@@ -623,7 +623,7 @@ _citation_author.ordinal
         self.assertEqual(w._repr(True), 'YES')
         # Test handling of 'long' type
         if sys.version_info[0] == 2:
-            self.assertEqual(w._repr(42L), '42')
+            self.assertEqual(w._repr(long(42)), '42')
 
     def test_cif_entities(self):
         """Test _EntityMapper class"""
